Split tests/test_resolvers.py into a package mirroring tests/test_plugins/
One test file per resolver (``test_venv.py``, ``test_pyproject.py``,
``test_uv_workspace.py``), one for ``exported_roots`` backend dispatch
(``test_exports.py``), and one for shared bits like ``merge_paths`` and
``load_resolver`` (``test_core.py``). The end-to-end ``tests/``-collision
regression test stays alongside the unit ``UvWorkspaceResolver`` tests
since it exercises the same resolver path end-to-end.

No test changes; just file moves.
